Replace Pattern3 with crfsuite
First pass on parsing the recipe ingredients.
Using crfsuite to parse the ingredients and using Pint to
recognize more easily the different units.

Still need to make the code more robust and refactor many things.
This is a work in progress.

Only problem: the output is still not working properly, as
described in issue #18.